Update main.css
Started the process change the webpage layout. started by adding float to the nav, main, and aside, and put a clear for the footer. adjusted the width to make everything fit, as well as added some padding, and font size changes in order to make it look similar to the photo.